requirements.txt
pillow
astropy
numpy
flask
flask-cors
matplotlib
ipython
-e file:///Users/mdaily/Documents/fits2image/
